requirements.txt
crewai==0.28.8
crewai_tools==0.1.6
python-dotenv==1.0.0
beautifulsoup4==4.13.3
langchain>=0.0.352
langchain-openai>=0.0.5